use workspace dns name instead of workspace name
Prior to this change, we used workspace name as host entry because it
is more human readable than using workspace dns name. This is a
problem for emacs TRAMP connections because of the `/` character in
some of our project names.

This change sets the Host parameter of the user's ssh config to the
workspace's dns name to support connections to a workspace from emacs
TRAMP